NOBUG: don't update existing passes in case of id collision
Because ids are randomly generated and `putItem` will update if the key
exists, if a collision does happen, the old pass with that duplicated
id will be overwritten with data from the new pass.

Collisions are unlikely to happen (1/100 000 000 initially I
think), but become more likely with each pass issued, so it's
possible it will happen a few times during the season.

Note that if this does happen the user will get a generic "Operation
Failed" error, but a retry will probably succeed.